FEAT - Refactor the Package Locations for Components and Services
